Reader: Move featured image under title in default card
Moved featured image & video under the title in default card for wide viewports.

Because the title, feature-image and excerpt order are changed, that order carries over to narrow viewports where the elements are stacked. Note that flex order can’t be used to move the title under the image (as currently is) because the image and excerpt are siblings one div down the DOM.

#9557