Media: allow embed-link style URLs for upload-via-url
Many screen-shotting services, like cloudup or droplr, give you a URL that takes you to a webpage that features that image... but you can get a direct link to the image file if you just add a "+" to the end of the URL (example: http://cld.wthms.co/Hl06+). The problem is, if you use that URL with "add via url," at least in Calypso, we see that the filename doesn't end in a known media extension and we disallow it saying "the file type is not supported." This branch removes the client restriction. We would still need to handle the extension restriction on the server.