Fix layout of the LanguagePickerModal dialog in Firefox
When the list of languages in the LanguagePickerModal is long and is supposed
to scroll, it doesn't scroll and overflows instead.

Fixed by adding an "overflow:hidden" property to the .dialog__content style.
This will change how the minimum content height of the element is evaluated
and will allow it to shrink properly.